Remove database logic from aggregator.js
