+++ .gitignore
@@ -1,4 +1,4 @@
 /node_modules
 /src/seeder
 .DS_Store
-.env.development.local
+.env
