+++ package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "@types/mongoose": "^5.11.97",
     "cors": "^2.8.5",
+    "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "joi": "^17.4.2",
     "mongoose": "^6.0.3",
