+++ src/schemas.js
@@ -1,8 +1,18 @@
 const mongoose = require('mongoose')
 
+// Schemas
+
 const ingredientSchema = new mongoose.Schema({
   name: String,
   recipe_link: String
 })
 
-module.exports = { ingredientSchema }
+// Models
+
+const BaseLayer = mongoose.model('Base Layer', ingredientSchema)
+const Condiment = mongoose.model('Condiment', ingredientSchema)
+const Mixin = mongoose.model('Mixin', ingredientSchema)
+const Seasoning = mongoose.model('Seasoning', ingredientSchema)
+const Shell = mongoose.model('Shell', ingredientSchema)
+
+module.exports = { BaseLayer, Condiment, Mixin, Seasoning, Shell }
