+++ src/server.js
@@ -1,9 +1,16 @@
+require('dotenv').config()
+
 const express = require('express')
 const cors = require('cors')
+const mongoose = require('mongoose')
 const generate = require('./routes/generate')
 
 const app = express()
 const port = process.env.PORT || 3001
+
+mongoose.connect(`${process.env.DB_URI}/tiktaco`)
+  .then(() => console.log('Connected to MongoDB...'))
+  .catch((err) => console.log('Could not connect to MongoDB', err))
 
 app.use(cors({
   origin:
